+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 - Unreleased
     - Added Column Vector
-    - Implemented Singular Value Decomposition
     - Implemented Eigenvalue decomposition
+    - Added solve system of linear equations
     - Integration with 3rd party JAMA libray
     - Fixed variance covariance calculation